# docs/swagger.yaml
definitions:
...
  payment.CreateOrderRequest:
    properties:
      amount:
        type: number
      order_name:
        maxLength: 64
        type: string
      remark:
        maxLength: 200
        type: string
    required:
    - amount
    - order_name
    type: object
  payment.PayOrderRequest:
    properties:
      order_no:
        type: string
      pay_key:
        maxLength: 10
        type: string
    required:
    - order_no
    - pay_key
    type: object
...
  util.ResponseAny:
    properties:
      data: {}
      error_msg:
        example: ""
        type: string
    type: object
info:
  contact: {}
  title: LINUX DO PAY
  version: 0.1.0
paths:
...
  /api/v1/merchant/payment:
    post:
      consumes:
      - application/json
      parameters:
      - description: 支付订单请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payment.PayOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - payment
  /api/v1/merchant/payment/order:
    get:
      consumes:
      - application/json
      parameters:
      - description: 订单号
        in: query
        name: order_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - payment
  /api/v1/merchant/payment/orders:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer {ClientID}:{ClientSecret}
        in: header
        name: Authorization
        required: true
        type: string
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payment.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - payment
...
swagger: "2.0"